# templating: fix templateSpecs.Less()