Using the length of the slice returned by strings.Split() to count the number of separators in a string is incredibly wasteful and makes for expensive sort operations.  This commit fixes that.

```sh
$ benchcmp base.txt new.txt
benchmark                    old ns/op     new ns/op     delta
BenchmarkTemplateLess-16     180           12.5          -93.06%

benchmark                    old allocs     new allocs     delta
BenchmarkTemplateLess-16     2              0              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkTemplateLess-16     192           0             -100.00%
```

### Required for all PRs:

- [ x ] Signed [CLA](https://influxdata.com/community/cla/).
- [ x ] Associated README.md updated.
N/A
- [ x ] Has appropriate unit tests.
A benchmark was added to identify regressions.